pyproject.toml
[tool.poetry]
name = "do-know-fastapi"
version = "0.1.0"
description = ""
authors = ["LucasDatilioCarderelli <llucascarderelli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.112.1"}
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '--disable-warnings'

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'fastapi dev do_know_fastapi/app.py'
lint = 'ruff check . & ruff check . --diff'
format = 'ruff check . --fix & ruff format .'

pre_test = 'task lint'
test = 'pytest --cov=do_know_fastapi -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
